- fix bug: change in 5.6.2 cause path transform fails with index.html in path.
when we are trying to fix cache resolving issue, we thought cache isn't
working because we didn't store normalized path in class path. So, we
move the path normalization code before setting class path.

however, we should keep the class path as is since transform uses it to
determine the base url; for example, base dir of `/apply/index.html` is
`/apply/`, but base dir of `/apply` is `/`.

instead, the real reason that manually set class with `index.html` in
path is because it didn't use normalized path as cache key, which is
inconsistent with `_get`, so we revert the change in `_get` and
normalize path in `set`, and the issue is solved.

 - bump verion